Move WritePreparedTxnDB to its own file
