+++ TARGETS
@@ -256,6 +256,7 @@ cpp_library(
       "utilities/transactions/transaction_lock_mgr.cc",
       "utilities/transactions/transaction_util.cc",
       "utilities/transactions/write_prepared_txn.cc",
+      "utilities/transactions/write_prepared_txn_db.cc",
       "utilities/ttl/db_ttl_impl.cc",
       "utilities/write_batch_with_index/write_batch_with_index.cc",
       "utilities/write_batch_with_index/write_batch_with_index_internal.cc",
