+++ utilities/transactions/write_prepared_transaction_test.cc
@@ -36,6 +36,7 @@
 #include "utilities/merge_operators.h"
 #include "utilities/merge_operators/string_append/stringappend.h"
 #include "utilities/transactions/pessimistic_transaction_db.h"
+#include "utilities/transactions/write_prepared_txn_db.h"
 
 #include "port/port.h"
